refactor(frontend): split playtest room lighting runtime
