+++ engine/examples/editor-playtest/src/room_lighting_runtime.rs
@@ -1,0 +1,282 @@
+use super::*;
+
+/// Walk `room.material_first..material_first + material_count`,
+/// resolve each material's texture asset, and build a
+/// TextureMaterial in `out` indexed by `local_slot`. Each
+/// texture asset is uploaded at most once across the program
+/// lifetime -- the residency manager + VRAM_SLOTS tracks who's
+/// already up.
+///
+/// Returns the highest `local_slot + 1` so the caller knows the
+/// in-use prefix length.
+pub(super) fn build_room_materials(
+    room: &LevelRoomRecord,
+    out: &mut [Option<WorldRenderMaterial>; MAX_ROOM_MATERIALS],
+) -> usize {
+    let first = room.material_first.to_usize();
+    let count = room.material_count as usize;
+    let slice: &[LevelMaterialRecord] = &MATERIALS[first..first + count];
+
+    let mut max_slot: usize = 0;
+    for material in slice {
+        let slot = material.local_slot.to_usize();
+        if slot >= MAX_ROOM_MATERIALS {
+            continue;
+        }
+        if slot + 1 > max_slot {
+            max_slot = slot + 1;
+        }
+        let Some(asset) = find_asset_of_kind(ASSETS, material.texture_asset, AssetKind::Texture)
+        else {
+            continue;
+        };
+        let Some(slot_record) = ensure_texture_uploaded(asset.id, asset.bytes) else {
+            continue;
+        };
+        let texture = TextureMaterial::opaque(
+            slot_record.clut_word,
+            slot_record.tpage_word,
+            rgb_tuple(material.tint_rgb),
+        )
+        .with_texture_window(slot_record.texture_window);
+        let render_material = match material.sidedness() {
+            LevelMaterialSidedness::Front => WorldRenderMaterial::front(texture),
+            LevelMaterialSidedness::Back => WorldRenderMaterial::back(texture),
+            LevelMaterialSidedness::Both => WorldRenderMaterial::both(texture),
+        }
+        .with_texture_size(
+            vram_slot_texture_size_u8(slot_record.texture_width),
+            vram_slot_texture_size_u8(slot_record.texture_height),
+        );
+        out[slot] = Some(render_material);
+    }
+    max_slot
+}
+
+#[derive(Copy, Clone)]
+pub(super) struct RuntimeRoomLighting {
+    pub(super) room_index: RoomIndex,
+    pub(super) ambient: Rgb8,
+    pub(super) camera: WorldCamera,
+    pub(super) fog_enabled: bool,
+    pub(super) fog_rgb: Rgb8,
+    pub(super) fog_near: i32,
+    pub(super) fog_far: i32,
+}
+
+impl RuntimeRoomLighting {
+    pub(super) fn shade_model_material(
+        &self,
+        point: WorldVertex,
+        material: TextureMaterial,
+    ) -> TextureMaterial {
+        material.with_tint(self.shade_tint_at(point, material.tint()))
+    }
+
+    pub(super) fn shade_tint_at(&self, point: RoomPoint, base: (u8, u8, u8)) -> (u8, u8, u8) {
+        let tint = psx_engine::shade_material_tint_with_lights(
+            MaterialTint::from_tuple(base),
+            point.to_array(),
+            self.ambient,
+            self.point_lights(),
+        )
+        .to_tuple();
+        if !self.fog_enabled || self.fog_far <= self.fog_near {
+            return tint;
+        }
+        let depth = self.camera.view_vertex(point).z;
+        self.apply_fog_at_depth(tint, depth)
+    }
+
+    pub(super) fn shade_tint_at_depth(
+        &self,
+        point: RoomPoint,
+        base: (u8, u8, u8),
+        fog_weight: i32,
+    ) -> (u8, u8, u8) {
+        let tint = psx_engine::shade_material_tint_with_lights(
+            MaterialTint::from_tuple(base),
+            point.to_array(),
+            self.ambient,
+            self.point_lights(),
+        )
+        .to_tuple();
+        self.apply_fog_weight(tint, fog_weight)
+    }
+
+    pub(super) fn apply_fog_at_depth(&self, tint: (u8, u8, u8), depth: i32) -> (u8, u8, u8) {
+        self.apply_fog_weight(tint, self.fog_weight_at_depth(depth))
+    }
+
+    pub(super) fn apply_fog_weight(&self, tint: (u8, u8, u8), weight: i32) -> (u8, u8, u8) {
+        apply_room_fog_weight(tint, self.fog_rgb, weight)
+    }
+
+    pub(super) fn fog_weight_at_depth(&self, depth: i32) -> i32 {
+        room_fog_weight(depth, self.fog_enabled, self.fog_near, self.fog_far)
+    }
+
+    fn point_lights(&self) -> impl Iterator<Item = PointLightSample> + '_ {
+        LIGHTS
+            .iter()
+            .filter(move |light| light.room == self.room_index)
+            .map(|light| {
+                PointLightSample::from_rgb_intensity(
+                    [light.x, light.y, light.z],
+                    light.radius as i32,
+                    Rgb8::from_array(light.color),
+                    Q8::from_raw_u16(light.intensity_q8),
+                )
+            })
+    }
+
+    pub(super) fn apply_vertex_fog(&self, rgb: (u8, u8, u8), vertex: WorldVertex) -> (u8, u8, u8) {
+        if !self.fog_enabled || self.fog_far <= self.fog_near {
+            return rgb;
+        }
+        let depth = self.camera.view_vertex(vertex).z;
+        self.apply_fog_at_depth(rgb, depth)
+    }
+
+    pub(super) fn apply_vertex_fog_weight(&self, rgb: (u8, u8, u8), weight: i32) -> (u8, u8, u8) {
+        self.apply_fog_weight(rgb, weight)
+    }
+}
+
+impl WorldSurfaceLighting for RuntimeRoomLighting {
+    fn shade(
+        &self,
+        sample: WorldSurfaceSample,
+        material: WorldRenderMaterial,
+    ) -> WorldRenderMaterial {
+        material.with_tint(self.shade_tint_at(sample.center, material.texture.tint()))
+    }
+
+    fn shade_vertex(
+        &self,
+        _sample: WorldSurfaceSample,
+        vertex: RoomPoint,
+        material: WorldRenderMaterial,
+    ) -> (u8, u8, u8) {
+        self.shade_tint_at(vertex, material.texture.tint())
+    }
+
+    fn shade_vertices(
+        &self,
+        sample: WorldSurfaceSample,
+        vertices: [WorldVertex; 4],
+        material: WorldRenderMaterial,
+    ) -> [(u8, u8, u8); 4] {
+        if let Some(vertex_rgb) = sample.baked_vertex_rgb {
+            if !self.fog_enabled || self.fog_far <= self.fog_near {
+                return vertex_rgb;
+            }
+            return [
+                self.apply_vertex_fog(vertex_rgb[0], vertices[0]),
+                self.apply_vertex_fog(vertex_rgb[1], vertices[1]),
+                self.apply_vertex_fog(vertex_rgb[2], vertices[2]),
+                self.apply_vertex_fog(vertex_rgb[3], vertices[3]),
+            ];
+        }
+        [
+            self.shade_vertex(sample, vertices[0], material),
+            self.shade_vertex(sample, vertices[1], material),
+            self.shade_vertex(sample, vertices[2], material),
+            self.shade_vertex(sample, vertices[3], material),
+        ]
+    }
+
+    fn shade_vertices_with_depths(
+        &self,
+        sample: WorldSurfaceSample,
+        vertices: [WorldVertex; 4],
+        depths: [i32; 4],
+        material: WorldRenderMaterial,
+    ) -> [(u8, u8, u8); 4] {
+        if let Some(vertex_rgb) = sample.baked_vertex_rgb {
+            if !self.fog_enabled || self.fog_far <= self.fog_near {
+                return vertex_rgb;
+            }
+            return [
+                self.apply_vertex_fog_weight(vertex_rgb[0], depths[0]),
+                self.apply_vertex_fog_weight(vertex_rgb[1], depths[1]),
+                self.apply_vertex_fog_weight(vertex_rgb[2], depths[2]),
+                self.apply_vertex_fog_weight(vertex_rgb[3], depths[3]),
+            ];
+        }
+        [
+            self.shade_tint_at_depth(vertices[0], material.texture.tint(), depths[0]),
+            self.shade_tint_at_depth(vertices[1], material.texture.tint(), depths[1]),
+            self.shade_tint_at_depth(vertices[2], material.texture.tint(), depths[2]),
+            self.shade_tint_at_depth(vertices[3], material.texture.tint(), depths[3]),
+        ]
+    }
+
+    fn shade_cached_baked_vertices(
+        &self,
+        sample: WorldSurfaceSample,
+        depths: Option<[i32; 4]>,
+        _material: WorldRenderMaterial,
+    ) -> Option<[(u8, u8, u8); 4]> {
+        let vertex_rgb = sample.baked_vertex_rgb?;
+        if !self.fog_enabled || self.fog_far <= self.fog_near {
+            return Some(vertex_rgb);
+        }
+        let depths = depths?;
+        Some([
+            self.apply_vertex_fog_weight(vertex_rgb[0], depths[0]),
+            self.apply_vertex_fog_weight(vertex_rgb[1], depths[1]),
+            self.apply_vertex_fog_weight(vertex_rgb[2], depths[2]),
+            self.apply_vertex_fog_weight(vertex_rgb[3], depths[3]),
+        ])
+    }
+
+    fn uses_vertex_depths(&self) -> bool {
+        self.fog_enabled && self.fog_far > self.fog_near
+    }
+
+    fn uses_direct_baked_vertex_rgb(&self) -> bool {
+        !self.fog_enabled || self.fog_far <= self.fog_near
+    }
+
+    fn prepare_vertex_depth(&self, depth: i32) -> i32 {
+        self.fog_weight_at_depth(depth)
+    }
+
+    fn needs_surface_sample_center(&self, sample_has_baked_rgb: bool) -> bool {
+        !sample_has_baked_rgb
+    }
+}
+
+#[inline(always)]
+fn room_fog_weight(depth: i32, enabled: bool, fog_near: i32, fog_far: i32) -> i32 {
+    if !enabled || fog_far <= fog_near || depth <= fog_near {
+        return 0;
+    }
+    (((depth - fog_near).saturating_mul(256)) / (fog_far - fog_near)).clamp(0, 256)
+}
+
+#[inline(always)]
+fn apply_room_fog_weight(tint: (u8, u8, u8), fog_rgb: Rgb8, weight: i32) -> (u8, u8, u8) {
+    if weight <= 0 {
+        return tint;
+    }
+    if weight >= 256 {
+        return (fog_rgb.r, fog_rgb.g, fog_rgb.b);
+    }
+    let keep = 256 - weight;
+    (
+        blend_channel(tint.0, fog_rgb.r, keep, weight),
+        blend_channel(tint.1, fog_rgb.g, keep, weight),
+        blend_channel(tint.2, fog_rgb.b, keep, weight),
+    )
+}
+
+#[inline(always)]
+fn blend_channel(src: u8, fog: u8, keep: i32, weight: i32) -> u8 {
+    (((src as i32) * keep + (fog as i32) * weight) >> 8) as u8
+}
+
+pub(super) const fn rgb_tuple(rgb: [u8; 3]) -> (u8, u8, u8) {
+    (rgb[0], rgb[1], rgb[2])
+}
